refactor: remove unused components and hooks, update imports
