--- src/app/auth/create-account/page.tsx
@@ -1,63 +1,0 @@
-'use client';
-
-import React, { useState } from 'react';
-import { useRouter } from 'next/navigation';
-import AuthForm from '@/components/auth/AuthForm';
-import { useAuthState } from '@/hooks/useAuthState';
-import { useCheckUserExists } from '@/lib/queries/auth';
-
-const CreateAccount: React.FC = () => {
-  const [localEmail, setLocalEmail] = useState('');
-  const [isLoading, setIsLoading] = useState(false);
-  const [error, setError] = useState('');
-
-  const router = useRouter();
-  const { setEmail } = useAuthState();
-
-  // React Query hook for checking user existence
-  const { refetch: checkUser, isLoading: isCheckingUser } = useCheckUserExists(localEmail, false); // disabled by default
-
-  const handleEmailSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!localEmail) return;
-
-    setIsLoading(true);
-    setError('');
-
-    try {
-      // Store email in sessionStorage
-      setEmail(localEmail);
-
-      // Use React Query to check if user exists
-      const result = await checkUser();
-      const userExists = result.data?.exists;
-
-      if (userExists) router.push('/auth/login/password');
-      else router.push('/auth/create-account/password');
-    } catch (error) {
-      console.error('Error checking user:', error);
-      setError('Something went wrong. Please try again.');
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <AuthForm
-      title="Create Account"
-      subtitle="Sign up for a new account"
-      email={localEmail}
-      setEmail={setLocalEmail}
-      onSubmit={handleEmailSubmit}
-      isLoading={isLoading || isCheckingUser}
-      error={error}
-      buttonText="Continue"
-      loadingText="Checking..."
-      footerText="Already have an account?"
-      footerLinkText="Sign In"
-      footerLinkHref="/auth/login"
-    />
-  );
-};
-
-export default CreateAccount;
