+++ src/app/dashboard/interview/result/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Page = () => {
+  return <div></div>;
+};
+
+export default Page;
