+++ src/components/dashboard/MenuItems.tsx
@@ -2,15 +2,19 @@ import {
   MdDashboard,
   MdFlashOn,
   MdVisibility,
-  MdSchool,
-  MdExplore,
-  MdAdd,
-  MdCheckCircle,
+  MdAccessTime,
+  MdArchive,
+  MdArticle,
+  MdAirplay,
+  // MdSchool,
+  // MdExplore,
+  // MdAdd,
+  // MdCheckCircle,
   MdBook,
-  MdLibraryBooks,
-  MdBookmark,
-  MdHighlight,
-  MdCloudUpload,
+  // MdLibraryBooks,
+  // MdBookmark,
+  // MdHighlight,
+  // MdCloudUpload,
   MdPerson,
   // MdAssessment,
   // MdQuiz,
@@ -21,11 +25,11 @@ import {
   MdTrackChanges,
   // MdReport,
   // MdEmojiEvents,
-  MdAndroid,
-  MdChat,
-  MdSchedule,
-  MdHelp,
-  MdRecommend,
+  // MdAndroid,
+  // MdChat,
+  // MdSchedule,
+  // MdHelp,
+  // MdRecommend,
 } from 'react-icons/md';
 import { SidebarConfig } from './Sidebar';
 
@@ -51,13 +55,13 @@ const menuItems: SidebarConfig = {
           href: '/dashboard',
           type: 'link',
         },
-        {
-          id: 'quick-actions',
-          name: 'Quick Actions',
-          icon: <MdFlashOn />,
-          href: '/dashboard/quick-actions',
-          type: 'link',
-        },
+        // {
+        //   id: 'quick-actions',
+        //   name: 'Quick Actions',
+        //   icon: <MdFlashOn />,
+        //   href: '/dashboard/quick-actions',
+        //   type: 'link',
+        // },
       ],
     },
     {
@@ -104,78 +108,107 @@ const menuItems: SidebarConfig = {
       ],
     },
     {
-      id: 'ai-mentor',
-      name: 'AI Mentor',
-      icon: <MdAndroid />,
+      id: 'interview',
+      name: 'Interview',
+      icon: <MdAirplay />,
       type: 'heading',
       children: [
         {
-          id: 'chat-mentor',
-          name: 'Chat with Mentor',
-          icon: <MdChat />,
-          href: '/dashboard/ai-chat',
+          id: 'schedule-interview',
+          name: 'Schedule Interview',
+          icon: <MdAccessTime />,
+          href: '/dashboard/interview/schedule',
           type: 'link',
         },
         {
-          id: 'study-planner',
-          name: 'Study Planner',
-          icon: <MdSchedule />,
-          href: '/dashboard/study-planner',
+          id: 'interview-results',
+          name: 'Interview Results',
+          icon: <MdArchive />,
+          href: '/dashboard/interview/result',
           type: 'link',
         },
         {
-          id: 'ask-questions',
-          name: 'Ask Questions',
-          icon: <MdHelp />,
-          href: '/dashboard/ask-questions',
-          type: 'link',
-        },
-        {
-          id: 'get-recommendations',
-          name: 'Get Recommendations',
-          icon: <MdRecommend />,
-          href: '/dashboard/recommendations',
+          id: 'practice-questions',
+          name: 'Practice Questions',
+          icon: <MdArticle />,
+          href: '/dashboard/interview/prepare',
           type: 'link',
         },
       ],
     },
-    {
-      id: 'learning-paths',
-      name: 'Learning Paths',
-      icon: <MdSchool />,
-      type: 'heading',
-      children: [
-        {
-          id: 'my-paths',
-          name: 'My Paths',
-          icon: <MdPerson />,
-          href: '/dashboard/my-paths',
-          type: 'link',
-        },
-        {
-          id: 'browse-paths',
-          name: 'Browse All Paths',
-          icon: <MdExplore />,
-          href: '/dashboard/browse-paths',
-          type: 'link',
-        },
-        {
-          id: 'create-path',
-          name: 'Create Custom Path',
-          icon: <MdAdd />,
-          href: '/dashboard/create-path',
-          type: 'link',
-        },
-        {
-          id: 'completed-paths',
-          name: 'Completed Paths',
-          icon: <MdCheckCircle />,
-          href: '/dashboard/completed-paths',
-          type: 'link',
-          badge: 'New',
-        },
-      ],
-    },
+    // {
+    //   id: 'ai-mentor',
+    //   name: 'AI Mentor',
+    //   icon: <MdAndroid />,
+    //   type: 'heading',
+    //   children: [
+    //     {
+    //       id: 'chat-mentor',
+    //       name: 'Chat with Mentor',
+    //       icon: <MdChat />,
+    //       href: '/dashboard/ai-chat',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'study-planner',
+    //       name: 'Study Planner',
+    //       icon: <MdSchedule />,
+    //       href: '/dashboard/study-planner',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'ask-questions',
+    //       name: 'Ask Questions',
+    //       icon: <MdHelp />,
+    //       href: '/dashboard/ask-questions',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'get-recommendations',
+    //       name: 'Get Recommendations',
+    //       icon: <MdRecommend />,
+    //       href: '/dashboard/recommendations',
+    //       type: 'link',
+    //     },
+    //   ],
+    // },
+    // {
+    //   id: 'learning-paths',
+    //   name: 'Learning Paths',
+    //   icon: <MdSchool />,
+    //   type: 'heading',
+    //   children: [
+    //     {
+    //       id: 'my-paths',
+    //       name: 'My Paths',
+    //       icon: <MdPerson />,
+    //       href: '/dashboard/my-paths',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'browse-paths',
+    //       name: 'Browse All Paths',
+    //       icon: <MdExplore />,
+    //       href: '/dashboard/browse-paths',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'create-path',
+    //       name: 'Create Custom Path',
+    //       icon: <MdAdd />,
+    //       href: '/dashboard/create-path',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'completed-paths',
+    //       name: 'Completed Paths',
+    //       icon: <MdCheckCircle />,
+    //       href: '/dashboard/completed-paths',
+    //       type: 'link',
+    //       badge: 'New',
+    //     },
+    //   ],
+    // },
     {
       id: 'resume-builder',
       name: 'Resume Builder',
@@ -198,49 +231,49 @@ const menuItems: SidebarConfig = {
         },
       ],
     },
-    {
-      id: 'resources',
-      name: 'Resources',
-      icon: <MdBook />,
-      type: 'heading',
-      children: [
-        {
-          id: 'my-library',
-          name: 'My Library',
-          icon: <MdLibraryBooks />,
-          href: '/dashboard/library',
-          type: 'link',
-        },
-        {
-          id: 'bookmarks',
-          name: 'Bookmarks',
-          icon: <MdBookmark />,
-          href: '/dashboard/bookmarks',
-          type: 'link',
-        },
-        {
-          id: 'notes-highlights',
-          name: 'Notes & Highlights',
-          icon: <MdHighlight />,
-          href: '/dashboard/notes',
-          type: 'link',
-        },
-        {
-          id: 'browse-resources',
-          name: 'Browse Resources',
-          icon: <MdExplore />,
-          href: '/dashboard/browse-resources',
-          type: 'link',
-        },
-        {
-          id: 'upload-content',
-          name: 'Upload Content',
-          icon: <MdCloudUpload />,
-          href: '/dashboard/upload',
-          type: 'link',
-        },
-      ],
-    },
+    // {
+    //   id: 'resources',
+    //   name: 'Resources',
+    //   icon: <MdBook />,
+    //   type: 'heading',
+    //   children: [
+    //     {
+    //       id: 'my-library',
+    //       name: 'My Library',
+    //       icon: <MdLibraryBooks />,
+    //       href: '/dashboard/library',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'bookmarks',
+    //       name: 'Bookmarks',
+    //       icon: <MdBookmark />,
+    //       href: '/dashboard/bookmarks',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'notes-highlights',
+    //       name: 'Notes & Highlights',
+    //       icon: <MdHighlight />,
+    //       href: '/dashboard/notes',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'browse-resources',
+    //       name: 'Browse Resources',
+    //       icon: <MdExplore />,
+    //       href: '/dashboard/browse-resources',
+    //       type: 'link',
+    //     },
+    //     {
+    //       id: 'upload-content',
+    //       name: 'Upload Content',
+    //       icon: <MdCloudUpload />,
+    //       href: '/dashboard/upload',
+    //       type: 'link',
+    //     },
+    //   ],
+    // },
     // {
     // 	id: "profile-assessment",
     // 	name: "Profile & Assessment",
