--- src/hooks/index.ts
@@ -1,2 +1,0 @@
-export { useAuth } from './useAuth';
-export { useAuthState } from './useAuthState';
